Move SimpleTooltip to its own file
